+++ sentry-android-core/src/main/java/io/sentry/android/core/AnrV2Integration.java
@@ -18,9 +18,11 @@ import io.sentry.SentryOptions;
 import io.sentry.android.core.cache.AndroidEnvelopeCache;
 import io.sentry.android.core.internal.threaddump.Lines;
 import io.sentry.android.core.internal.threaddump.ThreadDumpParser;
+import io.sentry.android.core.internal.util.NativeEventUtils;
 import io.sentry.hints.AbnormalExit;
 import io.sentry.hints.Backfillable;
 import io.sentry.hints.BlockingFlushHint;
+import io.sentry.protocol.ArtContext;
 import io.sentry.protocol.DebugImage;
 import io.sentry.protocol.DebugMeta;
 import io.sentry.protocol.Message;
@@ -32,7 +34,6 @@ import io.sentry.util.HintUtils;
 import io.sentry.util.Objects;
 import java.io.BufferedReader;
 import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
 import java.io.Closeable;
 import java.io.IOException;
 import java.io.InputStream;
@@ -173,6 +174,9 @@ public class AnrV2Integration implements Integration, Closeable {
           debugMeta.setImages(result.debugImages);
           event.setDebugMeta(debugMeta);
         }
+        if (result.artContext != null) {
+          event.getContexts().setArt(result.artContext);
+        }
       }
       event.setLevel(SentryLevel.FATAL);
       event.setTimestamp(DateUtils.getDateTime(anrTimestamp));
@@ -194,7 +198,7 @@ public class AnrV2Integration implements Integration, Closeable {
         if (trace == null) {
           return new ParseResult(ParseResult.Type.NO_DUMP);
         }
-        dump = getDumpBytes(trace);
+        dump = NativeEventUtils.readBytes(trace);
       } catch (Throwable e) {
         options.getLogger().log(SentryLevel.WARNING, "Failed to read ANR thread dump", e);
         return new ParseResult(ParseResult.Type.NO_DUMP);
@@ -209,6 +213,7 @@ public class AnrV2Integration implements Integration, Closeable {
 
         final @NotNull List<SentryThread> threads = threadDumpParser.getThreads();
         final @NotNull List<DebugImage> debugImages = threadDumpParser.getDebugImages();
+        final @Nullable ArtContext artContext = threadDumpParser.getArtContext();
 
         if (threads.isEmpty()) {
           // if the list is empty this means the system failed to capture a proper thread dump of
@@ -217,24 +222,10 @@ public class AnrV2Integration implements Integration, Closeable {
           // fall back to not reporting them
           return new ParseResult(ParseResult.Type.NO_DUMP);
         }
-        return new ParseResult(ParseResult.Type.DUMP, dump, threads, debugImages);
+        return new ParseResult(ParseResult.Type.DUMP, dump, threads, debugImages, artContext);
       } catch (Throwable e) {
         options.getLogger().log(SentryLevel.WARNING, "Failed to parse ANR thread dump", e);
         return new ParseResult(ParseResult.Type.ERROR, dump);
-      }
-    }
-
-    private byte[] getDumpBytes(final @NotNull InputStream trace) throws IOException {
-      try (final ByteArrayOutputStream buffer = new ByteArrayOutputStream()) {
-
-        int nRead;
-        final byte[] data = new byte[1024];
-
-        while ((nRead = trace.read(data, 0, data.length)) != -1) {
-          buffer.write(data, 0, nRead);
-        }
-
-        return buffer.toByteArray();
       }
     }
   }
@@ -300,15 +291,17 @@ public class AnrV2Integration implements Integration, Closeable {
     }
 
     final Type type;
-    final byte[] dump;
+    final @Nullable byte[] dump;
     final @Nullable List<SentryThread> threads;
     final @Nullable List<DebugImage> debugImages;
+    final @Nullable ArtContext artContext;
 
     ParseResult(final @NotNull Type type) {
       this.type = type;
       this.dump = null;
       this.threads = null;
       this.debugImages = null;
+      this.artContext = null;
     }
 
     ParseResult(final @NotNull Type type, final byte[] dump) {
@@ -316,17 +309,20 @@ public class AnrV2Integration implements Integration, Closeable {
       this.dump = dump;
       this.threads = null;
       this.debugImages = null;
+      this.artContext = null;
     }
 
     ParseResult(
         final @NotNull Type type,
         final byte[] dump,
         final @Nullable List<SentryThread> threads,
-        final @Nullable List<DebugImage> debugImages) {
+        final @Nullable List<DebugImage> debugImages,
+        final @Nullable ArtContext artContext) {
       this.type = type;
       this.dump = dump;
       this.threads = threads;
       this.debugImages = debugImages;
+      this.artContext = artContext;
     }
   }
 }